+++ R/key-range.R
@@ -84,15 +84,17 @@ key_range_auto <- function(sep = "[^[:alnum:]]+", reverse = FALSE, ...) {
   check_bool(reverse)
   force(sep)
   force(reverse)
-  dots <- label_args(...)
+  dots <- extra_args(...)
   call <- current_call()
-  function(scale, aesthetic = NULL) {
+  fun <- function(scale, aesthetic = NULL) {
     range_from_label(
       scale = scale, aesthetic = aesthetic,
       sep = sep, reverse = reverse, extra_args = dots,
       call = call
     )
   }
+  class(fun) <- union("key_range_auto_function", class(fun))
+  fun
 }
 
 #' @rdname key_range
@@ -100,7 +102,7 @@ key_range_auto <- function(sep = "[^[:alnum:]]+", reverse = FALSE, ...) {
 key_range_manual <- function(start, end, name = NULL, level = NULL, ...) {
   df <- data_frame0(
     start = start, end = end, .label = name, .level = level,
-    !!!label_args(...)
+    !!!extra_args(...)
   )
   check_columns(df, c("start", "end"))
   class(df) <- c("key_range", "key_guide", class(df))